+++ src/go/Makefile
