+++ src/go/cmd/init_command.go
