+++ src/go/hooks/local_dir.go
